fixed-authorize-delete-interview

+++ controllers/interviews.js
@@ -147,14 +147,17 @@ exports.updateInterview = async (req, res, next) => {
 exports.deleteInterview = async (req, res, next) => {
     try {
         const interview = await Interview.findById(req.params.id);
-
+        
         if (!interview) {
             return res.status(404).json({
                 success: false,
                 message: `No interview with the id of ${req.params.id}`
             });
         }
-
+        
+        if (interview.user.toString() !== req.user.id && req.user.role !== 'admin') {
+            return res.status(401).json({ success: false, message: 'Not authorized' });
+        }
         await interview.deleteOne();
 
         res.status(200).json({
